feat: add webpack + indexeddb
