+++ .gitignore
@@ -1,0 +1,3 @@
+public/sw.js*
+public/assets
+node_modules/
