--- js/main.js
@@ -1,215 +1,0 @@
-let restaurants, neighborhoods, cuisines;
-let newMap;
-let markers = [];
-
-/**
- * Fetch neighborhoods and cuisines as soon as the page is loaded.
- */
-document.addEventListener('DOMContentLoaded', event => {
-  initMap(); // added
-  fetchNeighborhoods();
-  fetchCuisines();
-});
-
-/**
- * Fetch all neighborhoods and set their HTML.
- */
-fetchNeighborhoods = () => {
-  DBHelper.fetchNeighborhoods((error, neighborhoods) => {
-    if (error) {
-      // Got an error
-      console.error(error);
-    } else {
-      self.neighborhoods = neighborhoods;
-      fillNeighborhoodsHTML();
-    }
-  });
-};
-
-/**
- * Set neighborhoods HTML.
- */
-fillNeighborhoodsHTML = (neighborhoods = self.neighborhoods) => {
-  const select = document.getElementById('neighborhoods-select');
-  neighborhoods.forEach(neighborhood => {
-    const option = document.createElement('option');
-    option.innerHTML = neighborhood;
-    option.value = neighborhood;
-    select.append(option);
-  });
-};
-
-/**
- * Fetch all cuisines and set their HTML.
- */
-fetchCuisines = () => {
-  DBHelper.fetchCuisines((error, cuisines) => {
-    if (error) {
-      // Got an error!
-      console.error(error);
-    } else {
-      self.cuisines = cuisines;
-      fillCuisinesHTML();
-    }
-  });
-};
-
-/**
- * Set cuisines HTML.
- */
-fillCuisinesHTML = (cuisines = self.cuisines) => {
-  const select = document.getElementById('cuisines-select');
-
-  cuisines.forEach(cuisine => {
-    const option = document.createElement('option');
-    option.innerHTML = cuisine;
-    option.value = cuisine;
-    select.append(option);
-  });
-};
-
-/**
- * Initialize leaflet map, called from HTML.
- */
-initMap = () => {
-  if (window.L) {
-    self.newMap = L.map('map', {
-      center: [40.722216, -73.987501],
-      zoom: 12,
-      scrollWheelZoom: false,
-    });
-    L.tileLayer(
-      'https://api.tiles.mapbox.com/v4/{id}/{z}/{x}/{y}.jpg70?access_token={mapboxToken}',
-      DBHelper.mapBoxData
-    ).addTo(self.newMap);
-  }
-
-  updateRestaurants();
-};
-
-/**
- * Update page and map for current restaurants.
- */
-updateRestaurants = () => {
-  const cSelect = document.getElementById('cuisines-select');
-  const nSelect = document.getElementById('neighborhoods-select');
-
-  const cIndex = cSelect.selectedIndex;
-  const nIndex = nSelect.selectedIndex;
-
-  const cuisine = cSelect[cIndex].value;
-  const neighborhood = nSelect[nIndex].value;
-
-  DBHelper.fetchRestaurantByCuisineAndNeighborhood(
-    cuisine,
-    neighborhood,
-    (error, restaurants) => {
-      if (error) {
-        // Got an error!
-        console.error(error);
-      } else {
-        resetRestaurants(restaurants);
-
-        const resultNumber = document.getElementById('filter-results');
-        resultNumber.innerHTML = `${restaurants.length} restaurants found`;
-
-        fillRestaurantsHTML();
-      }
-    }
-  );
-};
-
-/**
- * Clear current restaurants, their HTML and remove their map markers.
- */
-resetRestaurants = restaurants => {
-  // Remove all restaurants
-  self.restaurants = [];
-  const ul = document.getElementById('restaurants-list');
-  ul.innerHTML = '';
-
-  // Remove all map markers
-  if (self.markers) {
-    self.markers.forEach(marker => marker.remove());
-  }
-  self.markers = [];
-  self.restaurants = restaurants;
-};
-
-/**
- * Create all restaurants HTML and add them to the webpage.
- */
-fillRestaurantsHTML = (restaurants = self.restaurants) => {
-  const ul = document.getElementById('restaurants-list');
-  restaurants.forEach(restaurant => {
-    ul.append(createRestaurantHTML(restaurant));
-  });
-  addMarkersToMap();
-};
-
-/**
- * Create restaurant HTML.
- */
-createRestaurantHTML = restaurant => {
-  const li = document.createElement('li');
-
-  const image = document.createElement('img');
-  image.className = 'restaurant-img';
-  image.src = DBHelper.imageUrlForRestaurant(restaurant);
-  image.alt = restaurant.name;
-  image.setAttribute('aria-hidden', true);
-
-  const imageContainer = document.createElement('div');
-  imageContainer.className = 'restaurant-img-contianer';
-  imageContainer.append(image);
-
-  li.append(imageContainer);
-
-  const name = document.createElement('h3');
-  name.innerHTML = restaurant.name;
-
-  const nameContainer = document.createElement('div');
-  nameContainer.className = 'name-container';
-
-  const stripe = document.createElement('div');
-  stripe.className = 'stripe';
-
-  nameContainer.appendChild(name);
-  nameContainer.appendChild(stripe);
-
-  li.append(nameContainer);
-
-  const info = document.createElement('p');
-  info.innerHTML = `${restaurant.neighborhood} (${restaurant.cuisine_type})`;
-  li.append(info);
-
-  const address = document.createElement('p');
-  address.innerHTML = restaurant.address;
-  li.append(address);
-
-  const more = document.createElement('a');
-  more.innerHTML = `View Details <span class='visually-hidden'>of ${
-    restaurant.name
-  }</span>`;
-  more.href = DBHelper.urlForRestaurant(restaurant);
-  li.append(more);
-
-  return li;
-};
-
-/**
- * Add markers for current restaurants to the map.
- */
-addMarkersToMap = (restaurants = self.restaurants) => {
-  if (!window.L) return;
-
-  restaurants.forEach(restaurant => {
-    // Add marker to the map
-    const marker = DBHelper.mapMarkerForRestaurant(restaurant, self.newMap);
-    marker.on('click', onClick);
-    function onClick() {
-      window.location.href = marker.options.url;
-    }
-    self.markers.push(marker);
-  });
-};
