+++ package.json
@@ -1,0 +1,33 @@
+{
+  "name": "restaurant-reviews",
+  "author": "Jon Hargest",
+  "linkedIn": "https://www.linkedin.com/in/jonhargest/",
+  "private": true,
+  "version": "0.0.0",
+  "scripts": {
+    "start": "babel-node ./node_modules/webpack-dev-server/bin/webpack-dev-server --open",
+    "webpack": "babel-node ./node_modules/webpack/bin/webpack"
+  },
+  "main": "webpack.config.js",
+  "dependencies": {
+    "idb": "^3.0.2"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.3.3",
+    "@babel/node": "^7.2.2",
+    "@babel/preset-env": "^7.3.1",
+    "@babel/preset-react": "^7.0.0",
+    "babel-loader": "^8.0.5",
+    "clean-webpack-plugin": "^1.0.1",
+    "css-loader": "^2.1.0",
+    "file-loader": "^3.0.1",
+    "extract-text-webpack-plugin": "v4.0.0-alpha.0",
+    "node-sass": "^4.11.0",
+    "sass": "^1.17.2",
+    "sass-loader": "^7.1.0",
+    "style-loader": "^0.23.1",
+    "webpack": "^4.29.5",
+    "webpack-cli": "^3.2.3",
+    "webpack-dev-server": "^3.1.14"
+  }
+}
