+++ public/index.html
@@ -11,12 +11,13 @@
       rel="stylesheet"
       src="//normalize-css.googlecode.com/svn/trunk/normalize.css"
     />
-    <link rel="stylesheet" href="css/styles.css" />
+    <link rel="stylesheet" type="text/css" href="assets/app.min.css" />
     <link
       rel="stylesheet"
       href="https://unpkg.com/leaflet@1.3.1/dist/leaflet.css"
       integrity="sha512-Rksm5RenBEKSKFjgI3a41vrjkw4EVPlJ3+OiI65vTjIdo9brlAacEuKOiQ5OFh7cOI1bkDwLqdLw3Zg0cRJAAQ=="
       crossorigin=""
+      f
     />
   </head>
 
@@ -50,7 +51,7 @@
           <select
             id="neighborhoods-select"
             name="neighborhoods"
-            onchange="updateRestaurants()"
+            onchange="javascript: updateRestaurants()"
           >
             <option value="all">All Neighborhoods</option>
           </select>
@@ -72,11 +73,6 @@
     </main>
 
     <script
-      type="application/javascript"
-      charset="utf-8"
-      src="js/communal.js"
-    ></script>
-    <script
       src="https://unpkg.com/leaflet@1.3.1/dist/leaflet.js"
       integrity="sha512-/Nsx9X4HebavoBvEBuyp3I7od5tA0UzAxs+j83KgC8PU0kgB4XiK4Lfe4y4cgBtaRJQEIFCW+oC506aPT2L1zw=="
       crossorigin=""
@@ -84,12 +80,7 @@
     <script
       type="application/javascript"
       charset="utf-8"
-      src="js/dbhelper.js"
-    ></script>
-    <script
-      type="application/javascript"
-      charset="utf-8"
-      src="js/main.js"
+      src="/assets/index.bundle.js"
     ></script>
 
     <footer id="footer">
