+++ public/restaurant.html
@@ -8,7 +8,7 @@
       rel="stylesheet"
       src="//normalize-css.googlecode.com/svn/trunk/normalize.css"
     />
-    <link rel="stylesheet" href="css/styles.css" />
+    <link rel="stylesheet" href="assets/app.min.css" />
     <link
       rel="stylesheet"
       href="https://unpkg.com/leaflet@1.3.1/dist/leaflet.css"
@@ -59,16 +59,10 @@
     </footer>
 
     <script
-      type="application/javascript"
-      charset="utf-8"
-      src="js/communal.js"
-    ></script>
-    <script
       src="https://unpkg.com/leaflet@1.3.1/dist/leaflet.js"
       integrity="sha512-/Nsx9X4HebavoBvEBuyp3I7od5tA0UzAxs+j83KgC8PU0kgB4XiK4Lfe4y4cgBtaRJQEIFCW+oC506aPT2L1zw=="
       crossorigin=""
     ></script>
-    <script type="text/javascript" src="js/dbhelper.js"></script>
-    <script type="text/javascript" src="js/restaurant_info.js"></script>
+    <script type="text/javascript" src="assets/restaurant.bundle.js"></script>
   </body>
 </html>
