+++ src/shared/db.js
@@ -1,0 +1,9 @@
+import { openDb, deleteDb } from 'idb';
+
+export const dbPromise = () =>
+  openDb('mws-restaurants', 1, upgradeDB => {
+    switch (upgradeDB.oldVersion) {
+      case 0:
+        upgradeDB.createObjectStore('restaurants', { keyPath: 'id' });
+    }
+  });
