+++ src/shared/serviceworker.js
@@ -1,0 +1,6 @@
+const registerServiceWorker = () => {
+  if (!navigator.serviceWorker) return;
+  navigator.serviceWorker.register('/sw.js');
+};
+
+export default registerServiceWorker;
